requirements.txt
ipython==7.29.0
numpy==1.19.2
pandas==1.1.3
Pillow==8.4.0
plotly==5.1.0
streamlit==1.4.0
xlrd==1.2.0
openpyxl